+++ docs/CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to `database` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## 2017-12-12
+
+### Added
+- Ability to add custom options while connection using `ConnectionManager`
+
+### Removed
+- Useless `private` method `setConnection` in `ConnectionManager`
+
+### Fixed
+- Code readability
+
 ## 2017-11-22
 
 ### Deprecated